remove store from mobj position and direction
This breaks R1 and the svg view but maybe those can be fixed later. We are
seeing a ~10% improvement (from 107 tics to 117 in the perf test). fps also
increases about 10-15% (about ~60 with AI disabled in oversaturation map04 and now
about 72). Memory is also down about 10% (from 2.0GB to 1.8GB) but it's pretty
rough measure based on what the OS is reporting