remove light/flat/z value stores from sector
Mostly this is mechanically removing .val because the code was already
doing point-in-time comparison rather than subscribe. For a larger map
(cosmogenesis map05), we save about 100MB of RAM but performance isn't
really impacted which makes sense because the impact would be when we move
a sector and that isn't very often. We'll probably see a bigger change
when we remove store from linedef textures

r1 and the inspector are broken here.