remove stores from sidedef textures
This speeds up map load by about 30% (1300ms to 1000ms for cosmogenesis map05
and larger sunder maps are similar) but most maps won't make a meaningful
difference (even 300ms isn't that meaningful). Memory drops from about 1.7GB
down to 1.5GB.

Overall, it's an improvement but not as much as I had hoped. There are still
quite a few little functions created to update bits of state so there may
be room to improve