remove store from mobj.sector
Hardly seems worth it. We have only a couple of subscribers and the ui code
is cleaner with stores BUT we do save 100MB of memory usage on cosmogenesis
map05 and that is something (1.43GB from 1.54GB). Actually, it kind of makes
me want to remove other mobj stores (like health, sprite, etc.) because
using less memory is pretty meaningful! Also, most of the code really only
uses the point in time value and doesn't subscribe