simplify linedef update
We are basically always evaluating all parts of the line (upper, lower, middle)
so just update with a single function. We go from 1.4GB to about 1.3GB of memory
with cosmogenesis map05 so a 7% memory savings (which is good because less memory
== less gc == better experience)

With this change, we also create fewer empty upper/lower walls so for cosmogensis
map05 we go from 565K triangles to 508K. For Sunder MAP20, it's 483K (down from 566K).
Those are 10% drops which is good!