micro optimization
(1) Remove some foreach/find from collision detection critical path.
I think the benefit is mainly from less GC (AFAIK foreach creates an
iterator object which will impact GC). This is maybe a 3-5% speed boost
in perf test (620 tics to 640).
(2) use a map for thing type when creating new things. I modified
the perft test to nuts to see the benefit (because firing lots of guns
will spawn lots of things). We go from 2,100tics to 2,500 (it varies
so I'd call this a 15-20% win for maps that are spawning lots of
mobjs at the same time)