micro-optimize
(1) cleaner parans on blockmap radiusTrace
(2) cache 1/blockSize to avoid division
(3) cache for loop array limit to avoid extra property look up

In reality, there doesn't seem to be much perf impact. You could argue
it's not worth it but I would argue it doesn't change the readability
so why not? #3 in particular is a little odd and probably a smart js
engine would detect it anyway but why not help?