fix mobjs blocking crushers after being removed
Specifically blood was blocking a crusher but but it actually
made me realize we shouldn't add noblockmap objects into the sectormap

nuts.wad now doesn't drop into the 5fps territory... amazing! It's actually
a combination of things: this change, the missiles gliding on the floor change,
maybe the sound changes helped too?