fix aim slope
We stopped using slope as an angle because when we did that, sometimes
the vertical aim was off a little. Especially on shooting up/down at steep
angles.

This also affected rocket velocity but DOOM's physics are a little wrong on
that one anyway. Now we at least match the original DOOM