refactor mobj move/collision code
This separates all the missile-specific logic nicely. Unfortunately it
makes some members (pickup, hitFlat, hitC) public where they used to be
protected but that can probably be refactored and smoothed out later

Actually, with the code this way, we may be able to experiment with
not created a TraceParams object on each collision check. It's possible
JS engines optimize this away but it would be worth trying out later